% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R, R/message.R, R/print.R
\name{cat}
\alias{cat}
\alias{message}
\alias{quietR-outputs}
\alias{print}
\title{Output Functions}
\usage{
cat(...)

message(...)

print(...)
}
\arguments{
\item{...}{Arguments passed to masked output function}
}
\description{
Masked output functions (cat, message, print) that can be toggled with
[verbosity] functions. Instead of implementing your own verbose handling
system, use quietR.
}
\examples{
print("Hello World!")
disable_verbose()
print("Hello World!")
}
