% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R, R/message.R, R/print.R, R/verbosity.R
\name{enable_cat}
\alias{enable_cat}
\alias{disable_cat}
\alias{enable_message}
\alias{disable_message}
\alias{enable_print}
\alias{disable_print}
\alias{verbosity}
\alias{is_verbose}
\alias{enable_verbose}
\alias{disable_verbose}
\title{Getting and Setting Verbosity}
\usage{
enable_cat()

disable_cat()

enable_message()

disable_message()

enable_print()

disable_print()

is_verbose(func = c("print", "cat", "message"))

enable_verbose(func = c("print", "cat", "message"))

disable_verbose(func = c("print", "cat", "message"))
}
\arguments{
\item{func}{character - which function(s) to enable}
}
\value{
logical, whether or not to output specified function
}
\description{
Use the `is_verbose`, `enable_verbose` and `disable_verbose` to toggle
console output functions.
}
\examples{
# Returns FALSE:
is_verbose("print") #FALSE

enable_verbose()
# Now it returns TRUE:
is_verbose("print") #TRUE

disable_verbose()
# Now it returns FALSE:
is_verbose() #FALSE
is_verbose("print") #FALSE

enable_verbose("print")
print("Hi there!") # Outputs
cat("I won't output") # Doesn't output
}
